+++ concordium-sdk-examples/src/main/java/com/concordium/sdk/examples/GetBlockInfo.java
@@ -3,6 +3,7 @@ package com.concordium.sdk.examples;
 import com.concordium.sdk.ClientV2;
 import com.concordium.sdk.Connection;
 import com.concordium.sdk.Credentials;
+import com.concordium.sdk.exceptions.BlockNotFoundException;
 import com.concordium.sdk.exceptions.ClientInitializationException;
 import com.concordium.sdk.requests.BlockHashInput;
 import com.concordium.sdk.responses.blockinfo.BlockInfo;
@@ -24,7 +25,7 @@ public class GetBlockInfo implements Callable<Integer> {
     private String endpoint;
 
     @Override
-    public Integer call() throws ClientInitializationException, MalformedURLException {
+    public Integer call() throws ClientInitializationException, MalformedURLException, BlockNotFoundException {
         URL endpointUrl = new URL(this.endpoint);
         Connection connection = Connection.builder()
                 .host(endpointUrl.getHost())